+++ src/components/CookingInfo/index.tsx
@@ -2,7 +2,6 @@ import {
   Box,
   Heading,
   Flex,
-  Image,
   Input,
   Textarea,
   Text,
@@ -10,6 +9,7 @@ import {
 } from '@chakra-ui/react';
 import {PlusOutLineIcon} from 'components/Icon';
 import styled from '@emotion/styled';
+import ImageFileContainer from 'components/ImageFileBox';
 
 interface ICookingInfoPage {
   handleChange: (event: any) => void;
@@ -94,42 +94,6 @@ export default function CookingInfo({handleChange}: ICookingInfoPage) {
   );
 }
 
-function ImageFileContainer() {
-  return (
-    <Flex marginY={3} gap={2}>
-      <ImageFile key={1}></ImageFile>
-      <ImageFileInputWrapper>
-        <Input type={'file'} id={`fileInput`} hidden />
-        <label htmlFor={`fileInput`}>
-          <PlusOutLineIcon />
-        </label>
-      </ImageFileInputWrapper>
-    </Flex>
-  );
-}
-function ImageFile() {
-  return (
-    <ImageFileInputWrapper>
-      <Image
-        src="https://img.lovepik.com/element/40162/3354.png_860.png"
-        alt="Dan Abramov"
-        w={'100%'}
-        h={'100%'}
-      />
-    </ImageFileInputWrapper>
-  );
-}
-
-const ImageFileInputWrapper = styled(Box)`
-  border: 1px solid #b9b9b9;
-  width: 70px;
-  height: 70px;
-  border-radius: 3px;
-  display: inline-flex;
-  align-items: center;
-  border: 1px solid #b9b9b9;
-  justify-content: center;
-`;
 const CookingDesc = styled(Textarea)`
   border: none;
   border-top: 1px solid #b9b9b9;
