+++ src/components/RecipeStep/index.tsx
@@ -44,5 +44,3 @@ export default function RecipeStep({
     </Flex>
   );
 }
-
-
